# backend2/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: WordTrainer API
  version: 0.1.0
  description: >
    Backend2 API for the WordTrainer mobile app. Base path `/api/v1`, Sanctum bearer
    tokens, JSON, snake_case. This file is the contract — the Dart client is generated
    from it, so an endpoint not described here effectively does not exist.
    Errors are RFC 7807 `application/problem+json` with a stable machine `code`, except
    input validation which uses Laravel's 422 `{message, errors}` shape.
servers:
  - url: https://api.wordtrainer.app/api/v1
    description: Production (placeholder)
  - url: http://localhost:8001/api/v1
    description: Local (docker)
security:
  - bearerAuth: []
...
paths:
...
  /search:
    get:
      tags: [Search]
      summary: Search the terms that already exist (free, instant)
      description: >
        Exact and prefix matches over the term's own text and over its translations in the learner's
        language, ranked exact-term → exact-translation → prefix-term → prefix-translation. Costs
        nothing and touches no model, so it is safe to call on a debounced keystroke. Every hit says
        which of the CALLER'S OWN folders already hold it (`folders`), which is what lets the save
        button tell the truth. Deliberately not fuzzy: the question is «do we already have the word
        you typed», and a plausible near-miss would stop the learner generating the word they meant.

        `source`/`target` are the search pill, and they select the PAIR rather than a direction:
        «Похожие» are the same rows whichever way the learner is asking — `case` is an English term
        with a Russian translation either way — so both are searched and the translations shown are
        the pair's. Omit both for the learner's profile pair; an unserved pair is a 422.

        `translation` and `example_translation` are STRICTLY in the pair's support language: a term
        with no translation in it comes back with `null` there rather than with a gloss in a third
        language, and `POST /search/lookup` is what fills the gap (DECISIONS п. 146).
      parameters:
        - { name: q, in: query, required: true, schema: { type: string } }
        - { name: source, in: query, required: false, schema: { type: string, minLength: 2, maxLength: 2 } }
        - { name: target, in: query, required: false, schema: { type: string, minLength: 2, maxLength: 2 } }
        - name: taught_side
          in: query
          required: false
          schema: { type: string, enum: [source, target] }
          description: >-
            WHICH HALF of the pair is the language being STUDIED, said by the client instead of
            guessed by the server. A request carries a direction, not a pair of roles: `de → en` is
            either German with English support or English with German support, and the query string
            alone cannot tell them apart. Omit it and the server breaks the tie as before — the
            learner's profile, then the direction's `source` (DECISIONS п. 147). An unparseable
            value reads as «not given» on these two debounced GETs, because a typo must not put an
            error on a screen somebody is still typing into.
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 50, default: 20 } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/SearchHit' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /search/lookup:
    post:
      tags: [Search]
      summary: Look one word up with the model (paid, capped, cached forever)
      description: >
        ONE call to the cheap model for a word the database does not have. A POST the learner taps —
        never something a debounce fires — because it spends money.

        The order is cache → cap → call. A cached answer is served FREE and does not consult the cap
        at all: the answer is a fact about the word, not about who asked, so the second person to
        type it (and the same person tomorrow) pays nothing. Only a genuine miss counts against
        `SEARCH_LOOKUP_DAILY_CAP` (default 30/user/day).

        A spent cap is a **200**, not a 429: `limit_reached: true`, `lookup: null`. It is a normal
        answer the app has a screen for — the client shows the free `/search` results beside it.

        The term is NOT created. It appears only when the learner explicitly saves it via
        `/search/add`, which is what keeps the catalogue free of words nobody wanted.

        EITHER LANGUAGE, one card. `query` may be in the language being learned or in the learner's
        own; `lookup.text` is ALWAYS the former and `lookup.translation` always the latter, so
        `случай` and `occasion` produce the same word. They are two cache rows — the key is the
        query, and they really were two different questions — but they deduplicate to ONE term on
        save, because the term is the English word both questions were about.

        A query the model cannot place at all — keystrokes, a fragment — is also a **200**:
        `not_recognized: true`, `lookup: null`. The client's line is «не получилось распознать,
        проверьте написание», which is advice, and advice does not belong on an error path. The
        verdict is cached like a card (`asdfgh` will not become a word tomorrow) and counts against
        the day's cap like any other bought call.

        `source`/`target` are the same pill as on `GET /search/instant`, with the same rules: omit
        both for the learner's profile pair, and an unserved pair is a 422. The pair also decides
        what language the SAVED translation is written in — `/search/add` reads it off the cached
        lookup row rather than off the learner's profile.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:
                  type: string
                  minLength: 1
                  # The SAME ceiling `GET /search/instant` refuses at — services.generation
                  # .search_query_max_chars. A field that said «слишком длинно» while the button
                  # behind it bought the paragraph would be two products on one screen.
                  maxLength: 120
                source: { type: string, minLength: 2, maxLength: 2 }
                target: { type: string, minLength: 2, maxLength: 2 }
                taught_side:
                  type: string
                  enum: [source, target]
                  description: >-
                    WHICH HALF of the pair is the language being STUDIED, said by the client instead
                    of guessed by the server — the same optional field the two search GETs take. A
                    request carries a direction and not a pair of roles: `de → en` is either German
                    with English support or English with German support. Omitted, the server breaks
                    the tie as before (DECISIONS п. 147). Naming a side whose language this
                    deployment does not teach is a **422** `taught_side_not_taught`: this endpoint is
                    a deliberate tap and can afford to be told, and answering in the other role would
                    write the card in the wrong language invisibly.
      responses:
        '200':
          description: An answer, or an honest «not today»
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      lookup:
                        allOf: [{ $ref: '#/components/schemas/LookupCard' }]
                        nullable: true
                      not_recognized:
                        type: boolean
                        description: >
                          The model could not name a word for this query. `lookup` is null and this
                          is not an error — see the description above.
                      limit_reached: { type: boolean }
                      daily_cap: { type: integer }
                      used_today: { type: integer, description: paid lookups this UTC day }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422':
          description: >
            Either the input (`code: validation_error`) or the ANSWER (`code: lookup_refused`,
            `meta.reason` = `wrong_language` | `description_gives_away` | `empty_query` |
            `model_unavailable`). There is no repair pass: a lookup is a person watching a spinner,
            and a second call to fix the first would double both the wait and the cost.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /search/add:
    post:
      tags: [Search]
      summary: Save a search result — term, folder, POOL
      description: >
        Names either `lookup_id` (a freshly looked-up word: the cached answer becomes a term, with
        its description) or `term_id` (a word the database already had) — never both. `collection_id`
        omitted means «Сохранённые», created on the spot if this is the first save.

        It ENROLS the word. That obeys the pool rule rather than bending it: going looking for a
        word, typing it out and reading its card is as deliberate an act as «Учить это слово», and
        it is the same door. What the rule forbids is enrolling by side effect, which this is not.
        The enrichment станок is chained exactly as it is after a generation (`GENERATION_AUTO_ENRICH`).

        Idempotent throughout — the term dedups on its normalized text, the folder add is a no-op
        when the word is already there, the description is write-once, and re-enrolling an enrolled
        pair changes neither its rung nor its due date. `added`/`enrolled` say what actually happened.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                lookup_id: { type: string, minLength: 26, maxLength: 26 }
                term_id: { type: string, minLength: 26, maxLength: 26 }
                collection_id: { type: string, minLength: 26, maxLength: 26, nullable: true }
              oneOf:
                - required: [lookup_id]
                - required: [term_id]
      responses:
        '201':
          description: Saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      term_id: { type: string }
                      collection_id: { type: string }
                      collection_title: { type: string, description: the folder's CURRENT name — the learner may have renamed it }
                      collection_is_default: { type: boolean }
                      added: { type: boolean, description: false when the word was already in this folder }
                      enrolled: { type: boolean, description: false when the pair was already in the pool }
        '403':
          description: The folder is not the caller's (code `collection_not_editable`) — nothing is enrolled
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '404':
          description: No such cached lookup (code `lookup_not_found`)
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
    Unauthorized:
      description: Missing or invalid bearer token
...
    ValidationError:
      description: Input validation failed (Laravel shape)
      content:
        application/json:
          schema:
            type: object
            properties:
              message: { type: string }
              errors:
                type: object
                additionalProperties:
                  type: array
                  items: { type: string }
  schemas:
    Problem:
      type: object
      description: RFC 7807 with a stable machine `code`.
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail: { type: string }
        meta: { type: object, additionalProperties: true }
      required: [title, status, code]
...
    SearchHit:
      type: object
      description: One term the database already has, plus where it lives for THIS learner.
      properties:
        term_id: { type: string }
        text: { type: string }
        type: { type: string, enum: [word, phrase, idiom, phrasal_verb] }
        transcription: { type: string, nullable: true }
        translation: { type: string, nullable: true }
        description:
          type: string
          nullable: true
          description: >
            What the word means, written in the language BEING LEARNED. Null on the store
            catalogue — those terms have no description and are not being backfilled.
        example: { type: string, nullable: true }
        example_translation: { type: string, nullable: true }
        cefr: { type: string, nullable: true }
        folders:
          type: array
          description: The caller's OWN folders holding this term. Empty = not saved yet.
          items:
            type: object
            properties:
              id: { type: string }
              title: { type: string }
              is_default: { type: boolean }
    LookupCard:
      type: object
      description: A word the model looked up. Not a term yet — nothing is written until `/search/add`.
      properties:
        lookup_id: { type: string, description: the handle `/search/add` takes }
        text: { type: string }
        type: { type: string, enum: [word, phrase] }
        transcription: { type: string, nullable: true }
        translation: { type: string, nullable: true }
        description:
          type: string
          description: >
            One or two A2–B1 sentences in the language being learned, never containing the word
            itself — a description that names its headword answers the `description_match` card
            before it is asked, and the barrier refuses one that does.
        example: { type: string, nullable: true, description: 'contains the term verbatim, or is absent' }
        example_translation: { type: string, nullable: true }
        cefr: { type: string, nullable: true }
        fresh: { type: boolean, description: true when THIS call paid for the answer }
...